Relocate anvil command to command package from command/util package
