+++ paper/src/main/java/net/crystalixs/core/paper/bootstrap/PaperCommandBootstrap.java
@@ -5,7 +5,7 @@ import net.crystalixs.core.paper.CorePlugin;
 import net.crystalixs.core.paper.command.*;
 import net.crystalixs.core.paper.command.cloud.PaperCommandSource;
 import net.crystalixs.core.paper.command.cloud.PaperPlayerCommandSource;
-import net.crystalixs.core.paper.command.util.AnvilCommand;
+import net.crystalixs.core.paper.command.AnvilCommand;
 import net.crystalixs.core.paper.economy.DefaultEconomyService;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
